Move autocomplete routes to the find app
